CheckpointEntryPvp moved into separate files.
