+++ src/checkpointing/CheckpointEntryPvp.tpp
@@ -1,0 +1,96 @@
+/*
+ * CheckpointEntry.tpp
+ *
+ *  Created on Sep 27, 2016
+ *      Author: Pete Schultz
+ *  template implementations for CheckpointEntry class hierarchy.
+ *  Note that the .hpp includes this .tpp file at the end;
+ *  the .tpp file does not include the .hpp file.
+ */
+
+#include "io/fileio.hpp"
+#include "utils/PVAssert.hpp"
+#include "utils/PVLog.hpp"
+
+namespace PV {
+
+template <typename T>
+void CheckpointEntryPvp<T>::write(
+      std::string const &checkpointDirectory,
+      double simTime,
+      bool verifyWritesFlag) const {
+   PV_Stream *pvstream = nullptr;
+   if (getCommunicator()->commRank() == 0) {
+      std::string path = generatePath(checkpointDirectory, "pvp");
+      pvstream         = PV_fopen(path.c_str(), "w", verifyWritesFlag);
+      int *params      = pvp_set_nonspiking_act_params(
+            getCommunicator(), simTime, mLayerLoc, mDataType, 1 /*numbands*/);
+      pvAssert(params && params[1] == NUM_BIN_PARAMS);
+      int status = pvp_write_header(pvstream, getCommunicator(), params, params[1]);
+      pvErrorIf(
+            status != PV_SUCCESS,
+            "CheckpointEntryPvp::write unable to write header to to \"%s\"\n",
+            path.c_str());
+      free(params);
+   }
+   for (int b = 0; b < mLayerLoc->nbatch; b++) {
+      T *batchElementStart = calcBatchElementStart(b);
+      if (pvstream) {
+         PV_fwrite(&simTime, sizeof(double), (size_t)1, pvstream);
+      }
+      gatherActivity(
+            pvstream, getCommunicator(), 0 /*rootproc*/, batchElementStart, mLayerLoc, mExtended);
+   }
+   if (getCommunicator()->commRank() == 0) {
+      PV_fclose(pvstream);
+   }
+}
+
+template <typename T>
+void CheckpointEntryPvp<T>::read(std::string const &checkpointDirectory, double *simTimePtr) const {
+   std::string path    = generatePath(checkpointDirectory, "pvp");
+   PV_Stream *pvstream = pvp_open_read_file(path.c_str(), getCommunicator());
+   int rank            = getCommunicator()->commRank();
+   pvAssert((pvstream != nullptr && rank == 0) || (pvstream == nullptr && rank != 0));
+   int numParams = NUM_BIN_PARAMS;
+   int params[NUM_BIN_PARAMS];
+   int status = pvp_read_header(pvstream, getCommunicator(), params, &numParams);
+   if (status != PV_SUCCESS) {
+      read_header_err(path.c_str(), getCommunicator(), numParams, params);
+   }
+   MPI_Datatype *exchangeDatatypes = getCommunicator()->newDatatypes(mLayerLoc);
+   std::vector<MPI_Request> req{};
+   for (int b = 0; b < mLayerLoc->nbatch; b++) {
+      T *batchElementStart = calcBatchElementStart(b);
+      if (pvstream) {
+         PV_fread(simTimePtr, sizeof(double), (size_t)1, pvstream);
+      }
+      scatterActivity(
+            pvstream, getCommunicator(), 0 /*rootproc*/, batchElementStart, mLayerLoc, mExtended);
+      if (mExtended) {
+         getCommunicator()->exchange(batchElementStart, exchangeDatatypes, mLayerLoc, req);
+         getCommunicator()->wait(req);
+         // TODO: scattering should be aware of interprocess overlap region so that exchange call
+         // isn't necessary.
+      }
+   }
+   getCommunicator()->freeDatatypes(exchangeDatatypes);
+   MPI_Bcast(simTimePtr, 1, MPI_DOUBLE, 0, getCommunicator()->communicator());
+}
+
+template <typename T>
+T *CheckpointEntryPvp<T>::calcBatchElementStart(int batchElement) const {
+   int nx = mLayerLoc->nx;
+   int ny = mLayerLoc->ny;
+   if (mExtended) {
+      nx += mLayerLoc->halo.lt + mLayerLoc->halo.rt;
+      ny += mLayerLoc->halo.dn + mLayerLoc->halo.up;
+   }
+   return &mDataPointer[batchElement * nx * ny * mLayerLoc->nf];
+}
+
+template <typename T>
+void CheckpointEntryPvp<T>::remove(std::string const &checkpointDirectory) const {
+   deleteFile(checkpointDirectory, "pvp");
+}
+}
