+++ tests/CheckpointEntryTest/src/testPvpBatch.cpp
@@ -1,5 +1,5 @@
 #include "testPvpBatch.hpp"
-#include "checkpointing/CheckpointEntry.hpp"
+#include "checkpointing/CheckpointEntryPvp.hpp"
 #include "utils/conversions.h"
 
 void testPvpBatch(PV::Communicator *comm, std::string const &directory) {
